Move `InsightlyContacts`
